requirements.txt
openai
streamlit
pandas
plotly
python-dotenv
tqdm
scikit-learn
sentence-transformers
numpy
matplotlib
seaborn